pyproject.toml
[tool.poetry]
name = "plenkton-api"
version = "0.1.0"
description = ""
authors = ["Mikael Briolet <mbriolet.ma@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.73.0"
uvicorn = {extras = ["standard"], version = "^0.17.4"}
pytest = "^7.0.1"
requests = "^2.27.1"
black = "^22.1.0"
gunicorn = "^20.1.0"
httpx = "^0.22.0"
strawberry-graphql = "^0.97.0"
alembic = "^1.7.6"
psycopg2 = "^2.9.3"
ormar = {extras = ["postgresql"], version = "^0.10.24"}
python-slugify = "^6.1.0"
readtime = "^1.1.1"
auth0-python = "^3.20.0"
ipython = "^8.1.1"
jupyterlab = "^3.3.0"
python-jose = "^3.3.0"
pytest-asyncio = "^0.18.3"
doit = "^0.35.0"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
